Revisiones gráficas finales I
Ajuste de velocidades para un framerate fijo de 30fps con verticalsync.
Cambio en los textos de las rondas por preparados, listos, ya.
Corrección de un bug en la acumulación de ventanas ganadas y perdidas.
Corrección de los espaciados y encajes de los textos en windows y mac
os. Ronda + 1 en el texto del ganador.

+++ Mac OS/of_v0.9.3/thereminTest/src/vhpLevelMenu.h
@@ -94,8 +94,8 @@ class vhpLevelMenu {
         ofImage *                    keko;
         ofImage *                    bg;
     
-        int                         alpha;
-        int                         alpha_increment;
+        float                        alpha;
+        float                        alpha_increment;
     
         // fbo
         ofFbo                       fbo;
