+++ Mac OS/of_v0.9.3/thereminTest/src/vhpPlayerMenu.h
@@ -100,8 +100,8 @@ class vhpPlayerMenu {
         ofImage *                   bg;
     
     
-        int                         alpha;
-        int                         alpha_increment;
+        float                       alpha;
+        float                       alpha_increment;
     
         //loader
         vector<ofImage*>            loadingSilge;
