+++ Windows 7/thereminTouch/src/vhpStandby.h
@@ -85,8 +85,9 @@ class vhpStandby {
         ofImage                     bases;
         ofImage                     shadow;
         ofImage                     back;
-        int                         alpha;
-        int                         alpha_increment;
+    
+        float                       alpha;
+        float                       alpha_increment;
     
         //loader
         vector<ofImage*>            loadingSilge;
